requirements_LIVE.txt
chainer
opencv-python
Pillow
numpy